Update name and usages to remove new from title, old version deleted
