+++ java/compaction/compaction-job-creation/src/main/java/sleeper/compaction/job/creation/SendCompactionJobToSqs.java
@@ -22,7 +22,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import sleeper.compaction.core.job.CompactionJob;
-import sleeper.compaction.core.job.CompactionJobSerDeNew;
+import sleeper.compaction.core.job.CompactionJobSerDe;
 import sleeper.core.properties.instance.InstanceProperties;
 
 import java.io.IOException;
@@ -43,7 +43,7 @@ public class SendCompactionJobToSqs {
     }
 
     public void send(CompactionJob compactionJob) throws IOException {
-        String serialisedJobDefinition = new CompactionJobSerDeNew().toJson(compactionJob);
+        String serialisedJobDefinition = new CompactionJobSerDe().toJson(compactionJob);
         LOGGER.debug("Sending compaction job with id {} to SQS", compactionJob.getId());
         SendMessageRequest sendMessageRequest = new SendMessageRequest()
                 .withQueueUrl(instanceProperties.get(COMPACTION_JOB_QUEUE_URL))
